requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
configparser==3.7.3
cryptography==2.5
Django==1.11.17
django-haystack==2.8.1
django-phonenumber-field==2.2.0
enum34==1.1.6
et-xmlfile==1.0.1
future==0.17.1
idna==2.8
jdcal==1.4
lxml==4.2.5
mysqlclient==1.4.2.post1
oauthlib==2.1.0
openpyxl==2.5.12
phonenumbers==8.10.6
protobuf==3.6.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.1
Whoosh==2.7.4
xlrd==1.2.0
